feat(expo): migrate TimezoneSelectNative to /pickers/timezone route
Replace the RN Modal inside TimezoneSelectNative with a router formSheet that
uses the native headerSearchBarOptions for the search UI. The trigger is a
plain control (TimezoneSelect) that pushes /pickers/timezone and reads the
chosen value back from the zustand picker-result slot.
